CreateVM now works
Add OsType and ControlPanel data to sql scripts
change ID types from long to int
Add SQL to get spec by tier
Add SQL to get OS type ID.

+++ core/src/main/java/com/godaddy/vps4/vm/ControlPanelService.java
@@ -9,5 +9,5 @@ public interface ControlPanelService {
 
     List<String> listControlPanelTypes();
 
-    Long getControlPanelId(String controlPanel);
+    int getControlPanelId(String controlPanel);
 }
