+++ core/src/main/java/com/godaddy/vps4/vm/OsTypeService.java
@@ -2,7 +2,7 @@ package com.godaddy.vps4.vm;
 
 public interface OsTypeService {
 	
-	Long getOsTypeId(String osType);
+	int getOsTypeId(String osType);
 
 	
 }
