+++ core/src/main/java/com/godaddy/vps4/vm/jdbc/JdbcOsTypeService.java
@@ -19,14 +19,14 @@ public class JdbcOsTypeService implements OsTypeService {
     }
 
 	@Override
-	public Long getOsTypeId(String osType) {
+	public int getOsTypeId(String osType) {
 		return Sql.with(dataSource).exec("SELECT os_type_id FROM os_type WHERE name=?", this::osTypeIdMapper, osType);
 	}
 
-	public Long osTypeIdMapper(ResultSet rs) throws SQLException {
+	public int osTypeIdMapper(ResultSet rs) throws SQLException {
 		try{
 			rs.next();
-			return rs.getLong("os_type_id");
+			return rs.getInt("os_type_id");
 		} catch (SQLException sqlEx) {
 			throw new IllegalArgumentException("OS Type does not exist");
 		}
